core: EventStreamProxy
Adding throttle and debugName methods to EventStreamProxy.
Restricting access to abstract self member, on the reasoning that implementations can supply a public member implementation if they really want to, but otherwise EventStreamProxy doesn't need to force implementations to expose it.
Adding compile-time test to help ensure that future methods on EventStream are also proxied on EventStreamProxy.